MIPS/MIPS64: Update mips*_get_page_size functions.
There are multiple ways the page size could be extracted:
	- vmcore contains the PAGESIZE symbol.
	- Kernel configuration (if present) contains one
	  of the CONFIG_PAGE_SIZE_* configuration options.
	- Symbol following the swapper_pg_dir symbol (if
	  present, both) is page-aligned.

Currently, not all of them are covered by mips*_get_page_size
functions, so machdep->pagesize takes default values which may
not be correct.

Also, add MIPS and MIPS64 to the list of architectures that have
configurable page sizes in netdump.c.

Signed-off-by: Dimitrije Milosevic <dimitrije.milosevic@syrmia.com>